master - Removed Sweater.Android Project
Removed the Android binding project. It seems that embeddinator has a
bunch of problems at the moment and looks dead (6+ months and no
commits). Attempting to create an Android binding to the Sweater.Core
library failed when linking to Microsoft.Extensions.Logging
dependencies. It would be nice to revist this in the future, but for now
this isn't working.